build-python: List projects which are known to build explicitly
I've been experimenting with building python with VS2015, and I've
found it hard to verify the build does not regress due to the
oportunistic nature of the build script. This changes the script
to list the projects to build explicitly, which enables us to
check the build for errors.

I've verified that the new script produces the same set of files.

Change-Id: Ic841820db698c60f69264cd14b6bf51faac45a73