Prepare windows scripts for cmake 3.5
New cmake seems to be smarter (or dumber) and thinks it has found
a libxml package. However, the compilation of clang after that
fails, as the found package does not have the required include
header.  Since we don't need the xml functionality anyway, make
sure cmake does not even try looking for it.

Also, get rid of the PYTHON_INCLUDE_DIR define while I'm in there,
as cmake was telling us that it was unused for a while now.

Change-Id: I54b55000ba38b64771fb677436679a4ddb31cb2f