Switch darwin build scripts to cmake
The darwin and linux scripts seemed sufficiently similar now to
merge them into a single script, which I called build-posix.sh.
Unlike the linux build, the darwin build uses system clang for
compilation, as there does not seem to be a prebuilt host
toolchain capable of building lldb (but I guess that's not worse
than the previous state, when we were using system xcode as well).

Change-Id: Ic40d7fe3f1096abb11f3c1fa1756221b305ee537