Fix test script on devices which do not have the "env" command
Some devices indeed did not have the env command, introduced in
a810288. Do to LLDB test runner deficiencies (which I am fixing
upstream) this did not actually get reported as error, tests just
silently did nothing. This fixes the test script to not rely on
the "env" command. It uses the "interactive" mode of the shell, as
adb is pretty bad at passing quoted arguments through the shell
command line.

Tested to work on all devices connected to the linux buildbot.

Change-Id: Ic04048e59cd6051599ad3b500b91a35c7efcb3a7