Use incremental builds on the buildbots
Building lldb takes a significant percentage of our buildbot time
(cmake - >30%, android > 50%), which means we learn if a CL breaks
an android build only after about 2 hours. This CL should make
that significantly faster by using incremental builds.

Incremental builds are only used if the preceding build was
successful -- if it failed, we automatically fall back to a full
build. This should provide us with a reasonable safeguard from
cmake changes which leave the build directory in a inconsistent
state from which it is impossible to recover (although this hasn't
happened very often in the last two years).

Change-Id: I5ac9fd05aea071e70719511eaa1f9b8bf623f029