Set ARCHIVER and OBJCOPY when running local tests
This prepares the buildbots for D20386. It is necessary because
the logic for finding the toolchain files will not work when we
are using the just-compiled ToT clang. In all other cases the tool
names can be derived from the compiler, but here the toolchain
could be in a completely unrelated path.

Change-Id: Ib9d6af1c7a3096ec7aca0895cdb282cf8ccf7dd7