Make sure we don't upload huge test traces
In case of a faulty CL, we can have thousands of core files
generated. Uploading them is not a good idea (and they will not be
useful anyway).

Instead, when we detect a large trace size (currently the
threshold is 4GB, log files alone can ge up to 2GB), we delete
most of the core files (leaving just a couple of them for
analysis).

Bug: 31918201
Change-Id: Iaae407f30be1ae9ce5f1820e98e24455fba5dd9c