Extract configuration logic from buildserver scripts
This removes the manual build configuration logic from buildserver
scripts and puts if into a (series of) cmake cache files.

The idea behind this is that we can then reuse the same files for
configuring local builds, and on ab (for the latter I will need to come
up with a different way of specifying the host compilers).

Instead of the hacked-up toolchain file in the lldb tree, the build now
uses the cmake toolchain file present in the android ndk.

It also switches the build to use API level 9 headers and libraries to
build the 32 bit targets (previously we build using API 21 and then
employed some hacks to make sure it runs on older devices). I have
cleaned up lldb to not use any APIs not present on API 9, so the hacks
can be avoided now.

Change-Id: Iffcdeb263c0c4b16041065c581781a433df2455d