Clean up ab/ build-android script
This switches the build-android.sh script to use the same cmake cache
files as the upstream buildbots. The main difference is now in how
llvm-tblgen targets are built (on the buildbot they whatever is there on
the system, while on ab they use the aosp prebuilts). To make that work
I have added a cmake cache and a toolchain file for a local-linux build
(the next step will be swithcing linux build to use that).

This change comes with a one big caveat though - it breaks the aosp ab/
build, as we don't have a good ndk to use there - it uses an environment
variable instead. I don't think this is much of a problem though, as we
have never used the aosp build for anything. And in the internal repo, I
will point this directly at the checked-in NDK.

I have created a couple of symlink to the android cache files to make
the names align with what the build expects. A future work here will be
to remove the symlinks and converge to a single name for all android
targets.

This also stop building mips targets, as the studio will stop shipping
them.

Change-Id: Ibcd99cc80571a3440f03a006164db3d587ce738a