Switch android test scripts away from standalone toolchains
With https://reviews.llvm.org/D30410, lldb's test suite will begin
expecting the android compiler to reside in the android ndk, and not the
standalone toolchain. Adjust our test scripts to that.

I've tested that the bash changes work locally. The cmd changes were
eyeballed, and may require some tweaking in the field.

Change-Id: I8820d55aba31b727aa652916c53e4382fedff02c