Fix darwin build
Darwin build on ab was failing because the system compiler on the buildbots was
too old to understand some of the constructs in lldb. Take this as an
opportunity to update the compiler used, and also make the build more hermetic
-- it's still quite far from perfect, but we are now at least able to use the
same compiler and libc++ as the buildbots. We still pick up the system headers
and libraries from random locations on the buildbot, but this is the same thing
that the platform build does.

Change-Id: I0be4d28e82510a81c997ed7876dfd1451fd83302