Next attempt at fixing darwin build
The libc++ headers I tried to use turned out to be incompatible with the system
headers on the buildbots. Use libc++ system headers instead. This is probably a
good idea anyway, since we're linking agains system libc++. We could use the
prebuilt headers and prebuilt libc++ if/when we decide to ship our own
libc++.dylib.

Change-Id: I2f84e6d687325d7c953354a54637b62bdb3863d4