Revert all attempts to fix darwin build
I ran into an impassable problem. Very newer clangs (like the one in
prebuilts) are not compatible with very old libc++'s (like the ones on
ab bots).

The only way I see to get past this is to ship our own libc++, which I
am not ready to do yet. The original compile error which triggered all
of this seems fairly easy to work around on the lldb side, so I'll just
try to do that.

Change-Id: Iff217cac29c127bdc0fefaef2010cffe2e096a44