[prebuilts] Fix script for building swig 3.0.12
running autogen.sh will not work on the buildbots, as they don't have
necessary autoconf (nor bison)  tools installed. Instead, I run those
locally and patch them in when needed (emulating a build from a
distribution tarball, which does not require these tools). autoconf and
bison generate text files, so checking these in is not too gross, and it
still enables us to build swig binaries on the buildbots.

Change-Id: Iace75e72217b463a028feeca3c320b20b790ce34