Make lldb-server to be dynamic, position independent  and use LLVM libc++
Since LLDB droped support of Android versions before API level 16,
lldb-server can
depend on dynamic libraries now. Before that lldb-server had to be
static in
order to work on Androids that (a) don't support PIE or (b) require PIE.
Now lldb-server is a proper position independent executable.

It also allows LLDB to switch to LLVM libc++ instad of default GNU
libstdc++,
because we couldn't make static version of LLVM properly build with
libc++. In future NDKs libc++ is said to be default and most supported
version of STL.

Bug: 65386492
Change-Id: Iffd98dc3cd462e4c84d8998c2dc67d001e18ba28