Use killall to terminate previous instances of lldb-server on API>=26
The poor man's killall implementation we had until now broke down on API
26 because its "ps" implementation got more conforming and by
default displayed only processes in the current session. This resulted
in hilarious errors in tests, as we would delete the lldb-server
executable, but still keep it running and connect to it in the
subsequent test runs.

The fix would be to add a "-A" argument to ps, but unfortunately older
ps versions choke on that. So, I am adding a different, simpler
implementation for this API level instead.

Change-Id: I7eeb98642fea7463d2da386482ea952386c139ad